move build functions to data_series, add new meas attributes
